menu now works

+++ include/myStruct.h
@@ -8,10 +8,10 @@ using namespace std;
 struct accountNode {
     string accountName;
     string accountNumber;
-    long int balance;
+    long double balance;
     string accountPin;
     accountNode* next;
-    accountNode(string name, string number, long int balance, string pin) : accountName(name), accountNumber(number), balance(balance), accountPin(pin), next(nullptr) {}
+    accountNode(string name, string number, long double balance, string pin) : accountName(name), accountNumber(number), balance(balance), accountPin(pin), next(nullptr) {}
 };
 
 #endif
